Adding a unit test for module schema generation (#716)
Adding this over-assertive unit test for modules schema generation. Asserting on the exact schema that is currently generated for this fairly complex module, so that we have one more way of being notified when something about the schema generation changes. The schema generation is coupled to certain logic in the CFN server-side and we should be cautious when making changes.

The module fragment is taken from the official sample templates in the CFN documentation: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/sample-templates-services-us-west-2.html#w2ab1c35c58c13c27